Move daemon fixtures to "where" they're used
