+++ tests/integration/shell/test_syndic.py
@@ -10,6 +10,7 @@
 # Import python libs
 from __future__ import absolute_import
 import logging
+from collections import OrderedDict
 
 # Import Salt Testing libs
 from tests.support.case import ShellCase
@@ -23,12 +24,30 @@ import salt.utils.yaml
 import salt.utils.platform
 
 # Import 3rd-party libs
+import psutil
 import pytest
 
 log = logging.getLogger(__name__)
 
 
-@pytest.mark.usefixtures('session_salt_syndic')
+@pytest.fixture(scope='module', autouse=True)
+def session_salt_syndic(request, session_salt_master_of_masters, session_salt_syndic):
+    request.session.stats_processes.update(OrderedDict((
+        ('Salt Syndic Master', psutil.Process(session_salt_master_of_masters.pid)),
+        ('       Salt Syndic', psutil.Process(session_salt_syndic.pid)),
+    )).items())
+    yield session_salt_syndic
+    request.session.stats_processes.pop('Salt Syndic Master')
+    request.session.stats_processes.pop('       Salt Syndic')
+
+    # Stop daemons now(they would be stopped at the end of the test run session
+    for daemon in (session_salt_syndic, session_salt_master_of_masters):
+        try:
+            daemon.terminate()
+        except Exception as exc:  # pylint: disable=broad-except
+            log.warning('Failed to terminate daemon: %s', daemon.__class__.__name__)
+
+
 class SyndicTest(ShellCase, testprogram.TestProgramCase, ShellCaseCommonTestsMixin):
     '''
     Test the salt-syndic command
